services/validation-worker/requirements.txt
pika==1.3.*
psycopg2-binary==2.9.*
minio==7.*
python-dotenv==1.0.*
